requirements.txt
astroid==1.5.3
backports.functools-lru-cache==1.4
configparser==3.5.0
Django==1.11.1
django-bootstrap4==0.0.1
django-jenkins==0.110.0
enum34==1.1.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
# mysqlclient==1.3.10
pep8==1.7.0
pycrypto==2.6.1
pylint==1.9.3
pylint-django==0.10.0
pytz==2017.2
singledispatch==3.4.0.3
six==1.10.0
virtualenv==15.1.0
wrapt==1.10.10
django-webtest==1.9.2
coverage==4.5.2
-e git+https://github.com/rodrigoamaral/django-fullcalendar.git#egg=django-fullcalendar

